WIP commit: Working on operation #add_respondents_data.
Savon is configured with element_form_default: :qualified which makes
all elements be fully qualified with <namespace:tag-name>. Turns out
that QuestBack does not like the <Type> element representing an Enum to
have the default <wsdl:Type>. If I make a SOAP request with the Type
element without namespace (but all other elements with namespace) it
works. If I make a request where element_form_default is not configured
and Savon does not include namespaces every requests fails.

..I have to leave work soon, but I'll continue this later.