Completed support for #add_respondents_data.
I have to make a comment on #transform_hash_for_quest_back.
Yes, it is really hackish, but no I ain't gonna
clean it up now.

Also, using #add_respondents_data requires
a huge hash with correct attributes. A better solution
will be to create ourself som structure (classes)
which we can initiate instead of manually putting
the responsibility on the client of this API to
create exactly correct hashes.

This will have to wait for another time :-(